// src/app/shared/interfaces/Cita.ts
export interface Cita {
    idCronograma:               null;
    idCita:                     number;
    nuevoCliente:               boolean;
    foto:                       string;
    idTipoCita:                 number;
    idCliente:                  number;
    idTipoCliente:              number;
    idUsuario:                  number;
    idMaquina:                  number;
    idSede:                     number;
    idEstado:                   number;
    idUsuarioAtendidoPor:       null;
    usuarioAtendidoPor:         null;
    sede:                       string;
    idGenero:                   number;
    fiesta:                     boolean;
    zonasCorporales:            null;
    citaMensajeAvisos:          null;
    citaMensajeNotas:           null;
    citaMensajeDetalles:        null;
    fechaCita:                  Date;
    horaInicio:                 string;
    horaTermino:                string;
    duracion:                   number;
    nombres:                    string;
    apellidos:                  string;
    paciente:                   string;
    seudonimoPaciente:          string;
    numeroCita:                 string;
    codigoPaciente:             string;
    horaCita:                   string;
    numeroDocumentoIdentidad:   string;
    numeroHistoria:             string;
    numerosCelulares:           string;
    tipoCita:                   string;
    tipoCitaCorto:              string;
    colorTipoCita:              string;
    colorTextoTipoCita:         string;
    numeroSesion:               number;
    zonas:                      string[];
    zonaContatenadaCitaDetalle: null;
    citaAvisos:                 any[];
    citaDetalles:               CitaDetalle[];
    total:                      number;
    estado:                     string;
    colorEstado:                string;
    pagado:                     boolean;
    textoPagado:                string;
    colorPagado:                string;
    tipoCliente:                string;
    colorTipoCliente:           string;
    colorTextoTipoCliente:      null;
    fechaRegistra:              Date;
    usuarioRegistra:            string;
    usuario:                    string;
    idEstadoPendiente:          null;
    resumen:                    null;
    maquina:                    null;
    efectividad:                number;
    satisfaccion:               number;
    idMedioContacto:            null;
    otroMedioContacto:          null;
    idDescuento:                number;
    descuentoAplicaA:           null;
    cuponDescuento:             null;
    fechaEncuesta:              Date;
    idFichaAdmision:            any;
    idCitaAsignacion:           null;
    idUsuarioAsignado:          null;
    fechaConfirmacion:          null;
    idServicio:                 number;
    servicio:                   string;
    servicioColor:              string;
    genero:                     null;
    idPreferente:               number;
    numeroBox:                  number;
    idMaquinaMarca:             null;
    pagoFinal?:                 number;
    precioDePagoFinal?:         number;
}

export interface CitaDetalle {
    id:            number;
    idCita:        number;
    idUsuario:     number;
    detalle:       string;
    fechaRegistra: Date;
    destacado:     boolean;
}